+++ src/main/menus/tr/linux.json
@@ -1,0 +1,80 @@
+{
+  "menu": [
+    {
+      "label": "Dosya",
+      "id": "menu.file",
+      "submenu": [
+        { "label": "Motrix Hakkında", "id": "app.about", "command": "application:about", "command-before": "application:show,index"},
+        { "type": "separator" },
+        { "label": "Ayarlar...", "id": "app.preferences", "command": "application:preferences", "command-before": "application:show,index"},
+        { "label": "Güncellemeleri denetle...", "id": "app.check-for-updates", "command": "application:check-for-updates" },
+        { "label": "Motrix'i gizle", "id": "app.hide", "role": "hide" },
+        { "label": "Diğerlerini gizle", "id": "app.hide-others", "role": "hideothers" },
+        { "label": "Hepsini Göster", "id": "app.unhide", "role": "unhide" },
+        { "type": "separator" },
+        { "label": "Motrix'ten çık", "id": "app.quit", "role": "quit" }
+      ]
+    },
+    {
+      "label": "Görev",
+      "id": "menu.task",
+      "submenu": [
+        { "label": "Yeni Görev",  "id": "task.new-task", "command": "application:new-task", "command-arg": "uri", "command-after": "application:show,index"},
+        { "label": "Yeni BT Görevi", "id": "task.new-bt-task", "command": "application:new-bt-task", "command-arg": "torrent", "command-after": "application:show,index"},
+        { "type": "separator" },
+        { "label": "Görevi Durdur", "id": "task.pause-task", "command": "application:pause-task" },
+        { "label": "Görevi Sürdür", "id": "task.resume-task", "command": "application:resume-task" },
+        { "label": "Görevi Sil", "id": "task.delete-task", "command": "application:delete-task" },
+        { "label": "Görevi Yukarı Taşı", "id": "task.move-task-up", "command": "application:move-task-up" },
+        { "label": "Görevi Aşağı taşı", "id": "task.move-task-down", "command": "application:move-task-down" },
+        { "type": "separator" },
+        { "label": "Bütün Görevleri Durdur", "id": "task.pause-all-task", "command": "application:pause-all-task" },
+        { "label": "Bütün Görevleri Sürdür", "id": "task.resume-all-task", "command": "application:resume-all-task" },
+        { "type": "separator" },
+        { "label": "Son Görevleri Temizle", "id": "task.clear-recent-tasks", "command": "application:clear-recent-tasks" }
+      ]
+    },
+    {
+      "label": "Düzenle",
+      "id": "menu.edit",
+      "submenu": [
+        { "label": "Geri Al", "id": "edit.undo", "role": "undo" },
+        { "label": "Yinele", "id": "edit.redo", "role": "redo" },
+        { "type": "separator" },
+        { "label": "Kes", "id": "edit.cut", "role": "cut" },
+        { "label": "Kopyala", "id": "edit.copy", "role": "copy" },
+        { "label": "Yapıştır", "id": "edit.paste", "role": "paste" },
+        { "label": "Sil", "id": "edit.delete", "role": "delete" },
+        { "label": "Hepsini Seç", "id": "edit.select-all", "role": "selectall" }
+      ]
+    },
+    {
+      "label": "Pencere",
+      "role": "window",
+      "id": "menu.window",
+      "submenu": [
+        { "label": "Yeniden yükle", "id": "window.reload", "role": "reload" },
+        { "label": "Kapat", "id": "window.close", "role": "close" },
+        { "label": "Küçült", "id": "window.minimize", "role": "minimize" },
+        { "label": "Yakınlaştır", "id": "window.zoom", "role": "zoom" },
+        { "label": "Tam Ekran Yap", "id": "window.toggle-fullscreen", "role": "togglefullscreen" },
+        { "type": "separator" },
+        { "label": "Tümünü Öne Getir", "id": "window.front", "role": "front" }
+      ]
+    },
+    {
+      "label": "Yardım",
+      "role": "help",
+      "id": "menu.help",
+      "submenu": [
+        { "label": "Motrix Web Sitesi", "id": "help.official-website", "command": "help:official-website" },
+        { "label": "Kılavuz", "id": "help.manual", "command": "help:manual" },
+        { "label": "Sürüm Notları...", "id": "help.release-notes", "command": "help:release-notes" },
+        { "type": "separator" },
+        { "label": "Sorun bildir", "id": "help.report-problem", "command": "help:report-problem" },
+        { "type": "separator" },
+        { "label": "Geliştirici Aracına geçiş yap", "id": "help.toggle-dev-tools", "role": "toggledevtools" }
+      ]
+    }
+  ]
+}
