+++ src/shared/locales/all.js
@@ -3,7 +3,7 @@ import eleLocaleZhCN from 'element-ui/lib/locale/lang/zh-CN'
 import eleLocaleTr from 'element-ui/lib/locale/lang/tr-TR'
 import appLocaleEnUS from '@shared/locales/en-US'
 import appLocaleZhCN from '@shared/locales/zh-CN'
-import appLocaleTrTR from '@shared/locales/tr-TR'
+import appLocaleTr from '@shared/locales/tr'
 
 const resources = {
   'en-US': {
@@ -12,10 +12,10 @@ const resources = {
       ...appLocaleEnUS
     }
   },
-  'tr-TR': {
+  'tr': {
     translation: {
       ...eleLocaleTr,
-      ...appLocaleTrTR
+      ...appLocaleTr
     }
   },
   'zh-CN': {
