+++ src/shared/locales/index.js
@@ -6,7 +6,7 @@
 export const supportLanguages = [
   'en-US',
   'zh-CN',
-  'tr-TR'
+  'tr'
 ]
 
 /**
@@ -21,10 +21,6 @@ export function getLanguage (locale = 'en-US') {
 
   if (locale.startsWith('en')) {
     return 'en-US'
-  }
-
-  if (locale.startsWith('tr')) {
-    return 'tr-TR'
   }
 
   if (locale.startsWith('zh')) {
