+++ src/shared/locales/tr/about.js
