+++ src/shared/locales/tr/app.js
