+++ src/shared/locales/tr/edit.js
