+++ src/shared/locales/tr/index.js
