+++ src/shared/locales/tr/menu.js
