+++ src/shared/locales/tr/preferences.js
