+++ src/shared/locales/tr/subnav.js
