+++ src/shared/locales/tr/window.js
